fix build error and warnings
