+++ src/components/common/badges/StatusBadge.tsx
@@ -1,9 +1,10 @@
 /* eslint-disable react-refresh/only-export-components */
 // https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/25#issuecomment-1729071347
 
-import { Badge, getBadgeType } from '.';
+import { Badge } from '.';
 import { StatusIcon } from '../statusIcon';
 import { classNames } from '@/utils/commonUtils';
+import { getBadgeStatusType, statusTextColor } from '@/utils/statusUtils';
 
 const StatusBadge = ({ status }: { status: string }) => {
   return (
@@ -11,50 +12,12 @@ const StatusBadge = ({ status }: { status: string }) => {
       <div className='pr-2'>
         <StatusIcon
           status={status}
-          className={classNames('animate-none', statusTextColor(getBadgeType(status)))}
+          className={classNames('animate-none', statusTextColor(getBadgeStatusType(status)))}
         ></StatusIcon>
       </div>
       <span>{status}</span>
     </Badge>
   );
-};
-
-export const statusTextColor = (type: string) => {
-  switch (type) {
-    case 'success':
-      return 'text-green-700';
-    case 'fail':
-      return 'text-red-700';
-    case 'warning':
-      return 'text-yellow-700';
-    case 'abort':
-      return 'text-gray-700';
-    case 'unknown':
-      return 'text-gray-700';
-    case 'running':
-      return 'text-blue-700';
-    default:
-      return 'text-indigo-700';
-  }
-};
-
-export const statusBackgroundColor = (type: string) => {
-  switch (type) {
-    case 'success':
-      return 'bg-green-100';
-    case 'fail':
-      return 'bg-red-100';
-    case 'warning':
-      return 'bg-yellow-100';
-    case 'abort':
-      return 'bg-gray-100';
-    case 'unknown':
-      return 'bg-gray-100';
-    case 'running':
-      return 'bg-blue-100';
-    default:
-      return 'bg-indigo-100';
-  }
 };
 
 export default StatusBadge;
